Added build instructions for ixxat and peak driver on windows

+++ src/vscp/drivers/level1/ixxat_vc2/common/ixxobj.h
@@ -353,7 +353,7 @@ public:
 	/*!
 		CAN channel on board
 	*/
-	uint8_t m_channel;			
+	int m_channel;			
 	
 	/*!
 		IXXAT filter for outgoing messages
